Move some Uptane tests into their own files.
Hopefully this encourages parallelism.
