+++ tests/CMakeLists.txt
@@ -142,6 +142,12 @@ if(BUILD_OSTREE)
     add_aktualizr_test(NAME uptane SOURCES uptane_test.cc fake_ostree.cc
                        PROJECT_WORKING_DIRECTORY NO_VALGRIND)
 
+    add_aktualizr_test(NAME uptane_implicit SOURCES uptane_implicit_test.cc fake_ostree.cc
+                       PROJECT_WORKING_DIRECTORY)
+
+    add_aktualizr_test(NAME uptane_key SOURCES uptane_key_test.cc fake_ostree.cc
+                       PROJECT_WORKING_DIRECTORY)
+
     add_aktualizr_test(NAME ostree SOURCES ostree_test.cc ${PROJECT_SOURCE_DIR}/src/ostree.cc PROJECT_WORKING_DIRECTORY NO_VALGRIND)
 
     add_aktualizr_test(NAME tuf SOURCES tuf_test.cc PROJECT_WORKING_DIRECTORY)
